refactor: move requests logic from Professional page and components to Professional context
